requirements.txt
# ==========================================
# JARVIS PREMIUM REQUIREMENTS
# ==========================================
# Core AI & API Clients
groq==0.11.0
google-genai==0.8.0
together==1.2.0
tavily-python==0.5.0
chromadb==0.5.5
edge-tts==6.1.9
cartesia==1.0.0
# Google Gmail Integration
google-api-python-client==2.147.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
# Voice & Audio
speechrecognition==3.10.4
pvporcupine==3.0.4
pvrecorder==1.2.3
pyaudio==0.2.14
pygame==2.6.0
# GUI & System Tray
PyQt5==5.15.10
pystray==0.19.0
Pillow==10.4.0
rich==13.8.1
# Utilities
python-dotenv==1.0.1
requests==2.32.3
psutil==6.0.0
pywhatkit==5.4
markdown2==2.5.0
arxiv==2.1.3
twilio==9.3.2
# ==========================================